storage/database: split rbac_store.go by entity type (Issue #452)
The DatabaseRBACStore had grown to 1,387 lines with 4 independent entity
types (Permission, Role, Subject, RoleAssignment) implementing identical
CRUD patterns in a single file. This violated the package's per-store
convention already established by audit_store.go, config_store.go,
tenant_store.go, etc.

Each entity type is now isolated in its own file, making it easy to
locate and modify specific entity logic without scrolling through an
unrelated 1,300-line file.

Changes:
- rbac_store.go: reduced to 147 lines — struct, constructor, Initialize, Close, validateTenantAccess
- rbac_permissions.go: 208 lines — Permission CRUD (5 methods)
- rbac_roles.go: 259 lines — Role CRUD (5 methods)
- rbac_subjects.go: 272 lines — Subject CRUD (5 methods)
- rbac_assignments.go: 223 lines — RoleAssignment CRUD (4 methods)
- rbac_bulk.go: 193 lines — StoreBulkPermissions, StoreBulkRoles, StoreBulkSubjects
- rbac_queries.go: 157 lines — GetSubjectRoles, GetRolePermissions, GetSubjectAssignments

No public API changes. All existing tests in plugin_test.go pass without
modification. All files are under 300 lines.

Fixes #452
